+++ src/lib/Scenes/Fair/Screens/FairDetail.tsx
@@ -235,6 +235,7 @@ export const FairDetailContainer = createPaginationContainer(
           edges {
             cursor
             node {
+              id
               ...FairBoothPreview_show
             }
           }